+++ Installer/OS/parrot_os.py
@@ -2319,16 +2319,17 @@ programs_parrot_os_6_1.append(('nwdiag',
 packetdiag3 -h
 """,True,'Data'))
 
-# HamClock
-programs_parrot_os_6_1.append(('HamClock',
+# OpenHamClock
+programs_parrot_os_6_1.append(('OpenHamClock',
 """mkdir -p ~/Installed_by_FISSURE
 cd ~/Installed_by_FISSURE
-wget https://www.clearskyinstitute.com/ham/HamClock/ESPHamClock.zip
-unzip -q ESPHamClock.zip
-rm ESPHamClock.zip
-cd ESPHamClock
-make install hamclock-1600x960
-sudo make install hamclock-1600x960
+rm -rf hamclock
+if timeout 120 git clone --depth 1 https://github.com/openhamclock/hamclock.git; then
+    cd hamclock/ESPHamClock
+    make hamclock-1600x960 && sudo make install
+else
+    echo "OpenHamClock download failed or timed out."
+fi
 ########## Verify ##########
 ls /usr/local/bin/hamclock
 """,True,'Ham Radio'))