+++ Installer/OS/raspberry_pi_os.py
@@ -2326,16 +2326,17 @@ programs_raspberry_pi_os.append(('nwdiag',
 packetdiag -h
 """,True,'Data'))
 
-# HamClock
-programs_raspberry_pi_os.append(('HamClock',
+# OpenHamClock
+programs_raspberry_pi_os.append(('OpenHamClock',
 """mkdir -p ~/Installed_by_FISSURE
 cd ~/Installed_by_FISSURE
-wget https://www.clearskyinstitute.com/ham/HamClock/ESPHamClock.zip
-unzip -q ESPHamClock.zip
-rm ESPHamClock.zip
-cd ESPHamClock
-make install hamclock-1600x960
-sudo make install hamclock-1600x960
+rm -rf hamclock
+if timeout 120 git clone --depth 1 https://github.com/openhamclock/hamclock.git; then
+    cd hamclock/ESPHamClock
+    make hamclock-1600x960 && sudo make install
+else
+    echo "OpenHamClock download failed or timed out."
+fi
 ########## Verify ##########
 ls /usr/local/bin/hamclock
 """,True,'Ham Radio'))